Fixes #271 - AbstractGedLine complexity
Create another layer of abstraction to spread the complexity around.
