+++ gedbrowser-reader/src/main/java/org/schoellerfamily/gedbrowser/reader/AbstractGedLine.java
@@ -13,7 +13,7 @@ import org.schoellerfamily.gedbrowser.reader.AbstractGedObjectFactory.GedObjectF
  *
  * @author Dick Schoeller
  */
-public abstract class AbstractGedLine {
+public abstract class AbstractGedLine extends AbstractSingleGedLine {
     /** */
     private static final GedObjectFactory GOB_FACTORY = new GedObjectFactory();
 
@@ -23,43 +23,39 @@ public abstract class AbstractGedLine {
     /** */
     private final transient List<AbstractGedLine> children =
             new ArrayList<AbstractGedLine>();
-    /** */
-    private final transient int lineNumber;
-    /** */
-    private transient int level;
-    /** */
-    private transient String xref = "";
-    /** */
-    private transient String tag = "";
-    /** */
-    private transient String tail = "";
-    /** */
-    private transient GedObject gedObject;
 
     /**
      * @param parent The parent GedLine. Parentage is determined by the GEDCOM
      *               level number.
      */
     protected AbstractGedLine(final AbstractGedLine parent) {
+        super(generateLineNumber(parent));
         if (parent == null) {
             this.source = new NullSource();
-            this.lineNumber = 0;
         } else {
             this.source = parent.source;
-            this.lineNumber = parent.source.getMaxLineNumber();
         }
         this.setLevel(-1);
-        this.gedObject = null;
+    }
+
+    /**
+     * @param parent the parent
+     * @return the appropriate line number
+     */
+    private static int generateLineNumber(final AbstractGedLine parent) {
+        if (parent == null) {
+            return 0;
+        }
+        return parent.source.getMaxLineNumber();
     }
 
     /**
      * @param reader The buffered reader that we are getting data from.
      */
     protected AbstractGedLine(final BufferedReader reader) {
+        super(0);
         this.source = new ReaderSource(reader);
-        this.lineNumber = 0;
         this.setLevel(-1);
-        this.gedObject = null;
     }
 
     /**
@@ -67,11 +63,10 @@ public abstract class AbstractGedLine {
      */
     @SuppressWarnings({ "PMD.UseVarargs", "PMD.ArrayIsStoredDirectly" })
     protected AbstractGedLine(final String[] arraySource) {
+        super(0);
         this.source = new ArraySource(arraySource);
-        this.lineNumber = 0;
         this.setLevel(-1);
         this.setTag("ROOT");
-        this.gedObject = null;
     }
 
     /**
@@ -93,119 +88,11 @@ public abstract class AbstractGedLine {
     }
 
     /**
-     * @return the current GEDCOM level
-     */
-    public final int getLevel() {
-        return level;
-    }
-
-    /**
-     * @param level the GEDCOM level for this line
-     */
-    public final void setLevel(final int level) {
-        this.level = level;
-    }
-
-    /**
-     * @return the cross reference string for this line
-     */
-    public final String getXref() {
-        return xref;
-    }
-
-    /**
-     * @param xref the cross reference string for this line
-     */
-    public final void setXref(final String xref) {
-        this.xref = xref;
-    }
-
-    /**
-     * @return the GEDCOM tag on this line
-     */
-    public final String getTag() {
-        return tag;
-    }
-
-    /**
-     * @param tag the GEDCOM tag on this line
-     */
-    public final void setTag(final String tag) {
-        this.tag = tag;
-    }
-
-    /**
-     * @return whatever text follows the tag
-     */
-    public final String getTail() {
-        return tail;
-    }
-
-    /**
-     * @param tail the text following the tag
-     */
-    public final void setTail(final String tail) {
-        this.tail = tail;
-    }
-
-    /**
-     * Sets the GedObject associated with this line.
-     *
-     * @param gedObject the GedObject
-     */
-    public final void setGedObject(final GedObject gedObject) {
-        this.gedObject = gedObject;
-    }
-
-    /**
-     * Gets the GedObject associated with this line.
-     *
-     * @return the GedObject
-     */
-    public final GedObject getGedObject() {
-        return gedObject;
-    }
-
-    /**
-     * Create the GedObject for this line.
-     *
-     * @param parentLine the parent GedLine.
-     * @return the GedObject
-     */
-    public final GedObject createGedObject(final AbstractGedLine parentLine) {
-        GedObject parentGob;
-        if (parentLine == null) {
-            parentGob = null;
-        } else {
-            parentGob = parentLine.gedObject;
-        }
-
-        final GedObject gob = createGedObject(parentGob);
-        if (gob == null) {
-            return null;
-        }
-
-        setGedObject(gob);
-
-        createChildren(gob);
-
-        return gob;
-    }
-
-    /**
-     * Create a GedObject for this line.
-     *
-     * @param parentGedObj the parent GedObject
-     * @return the GedObject
-     */
-    protected abstract GedObject createGedObject(GedObject parentGedObj);
-
-    /**
      * Create the GedObjects for the child lines.
      *
      * @param subObject the GedObject whose children are to be created.
      */
-    protected final void createChildren(final GedObject subObject) {
+    public final void createChildren(final GedObject subObject) {
         for (final AbstractGedLine child : getChildren()) {
             final GedObject childGob = child.createGedObject(this);
             subObject.insert(childGob);
@@ -258,15 +145,6 @@ public abstract class AbstractGedLine {
      */
     public final List<AbstractGedLine> getChildren() {
         return children;
-    }
-
-    /**
-     * Get the line number.
-     *
-     * @return the line number.
-     */
-    public final int getLineNumber() {
-        return lineNumber;
     }
 
     /**
